:fire: Backend DELETE done,

+++ backend/model/DificuldadeModel.php
@@ -2,6 +2,7 @@
 
 namespace Model;
 
+use Exception;
 use Helper\Connection;
 use Helper\Response;
 use PDO;
@@ -80,13 +81,16 @@ class DificuldadeModel
     {
         try {
             $con = Connection::getConn();
-            $this->get(array('id' => $id))[0];
-            $stmt = $con->prepare("DELETE FROM tb_dificuldade WHERE idDificuldade = ?");
-            $stmt->bindValue(1, trim($id), PDO::PARAM_INT);
-            if ($stmt->execute()) {
-                return Response::success("Dificuldade id=`$id` deletada com sucesso");
-            }
-            return Response::error("Erro ao deletar dificuldade");
+            $data = json_decode($this->get(array('id' => $id)));
+            if ($data->status_code === 200) {
+                $stmt = $con->prepare("DELETE FROM tb_dificuldade WHERE idDificuldade = ?");
+                $stmt->bindValue(1, trim($id), PDO::PARAM_INT);
+                if ($stmt->execute()) {
+                    return Response::success("Dificuldade id=`$id` deletada com sucesso");
+                }
+                return Response::warning("Erro ao deletar dificuldade", 404);
+            };
+            return Response::warning("Dificuldade id=$id nao encontrada", 404);
         } catch (\Throwable $th) {
             return Response::error("Error: " . $th->getMessage());
         }
